Fold customizable into character
