+++ maelstrom/dataClasses/character.py
@@ -1,13 +1,13 @@
 # TODO fix this in #12 
 # from maelstrom.dataClasses.activeAbilities import TargetOption
 from maelstrom.gameplay.events import ActionRegister, UPDATE_EVENT
-from maelstrom.dataClasses.customizable import AbstractCustomizable
 from maelstrom.dataClasses.stat_classes import Stat
+from maelstrom.util.serialize import AbstractJsonSerialable
 from maelstrom.util.stringUtil import entab, lengthOfLongest
 
 _STATS = ("control", "resistance", "potency", "luck", "energy")
 
-class Character(AbstractCustomizable):
+class Character(AbstractJsonSerialable):
     """
     A Character is an entity within the game who has various stats and attributes.
     """
@@ -24,6 +24,8 @@ class Character(AbstractCustomizable):
         """
 
         super().__init__(**dict(kwargs, type="Character"))
+        self.name = kwargs["name"]
+
         self._max_hp = 100
 
         # don't prefix with underscore - breaks JSON!
@@ -31,20 +33,25 @@ class Character(AbstractCustomizable):
         self.level = kwargs.get("level", 1)
         self.xp = int(kwargs.get("xp", 0))
         self.actives = [active for active in kwargs["actives"]]
-
+        self.stats = {}
         for stat in _STATS:
-            self.addStat(Stat(stat, lambda base: 20.0 + float(base), kwargs.get("stats", {}).get(stat, 0)))
-        self.calcStats()
+            self._add_stat(Stat(stat, lambda base: 20.0 + float(base), kwargs.get("stats", {}).get(stat, 0)))
+        self._calc_stats()
         self.remaining_hp = self._max_hp
 
         self._event_listeners = ActionRegister()
 
         self.addSerializedAttributes(
+            "name",
             "element", # todo rm once I use templates for #6
             "xp",
             "level",
-            "actives"
+            "actives",
+            "stats"
         )
+
+    def _add_stat(self, stat):
+        self.stats[stat.name.lower()] = stat
 
     def init_for_battle(self):
         """
@@ -52,12 +59,23 @@ class Character(AbstractCustomizable):
         """
 
         self._event_listeners.clear()
-        self.calcStats()
+        self._calc_stats()
 
         # don't need to do anything with actives
 
         self.remaining_hp = self._max_hp
-        self.energy = int(self.getStatValue("energy") / 2.0)
+        self.energy = int(self.get_stat_value("energy") / 2.0)
+
+    def _calc_stats(self):
+        """
+        Calculates all this' stats
+        """
+        for stat in self.stats.values():
+            stat.reset_boosts()
+            stat.calc()
+
+    def get_stat_value(self, statName: str) -> float:
+        return self.stats[statName.lower()].get()
 
     def add_event_listener(self, enum_type, action):
         self._event_listeners.add_event_listener(enum_type, action)
@@ -72,7 +90,7 @@ class Character(AbstractCustomizable):
         return int((float(self.remaining_hp) / float(self._max_hp) * 100.0))
 
     def get_display_data(self) -> str:
-        self.calcStats()
+        self._calc_stats()
         ret = [
             f'{self.name} Lv. {self.level} {self.element}',
             entab(f'{self.xp} / {self.level * 10} XP')
@@ -81,7 +99,7 @@ class Character(AbstractCustomizable):
         ret.append("STATS:")
         width = lengthOfLongest(_STATS)
         for stat in _STATS:
-            ret.append(entab(f'{stat.ljust(width)}: {int(self.getStatValue(stat))}'))
+            ret.append(entab(f'{stat.ljust(width)}: {int(self.get_stat_value(stat))}'))
 
         ret.append("ACTIVES:")
         for active in self.actives:
@@ -111,7 +129,7 @@ class Character(AbstractCustomizable):
         Increase or lower stats in battle. Returns the boost this receives with its
         potency stat factored in
         """
-        mult = 1 + self.getStatValue("potency") / 100
+        mult = 1 + self.get_stat_value("potency") / 100
         boost = boost.copy()
         boost.amount *= mult
         self.stats[boost.stat_name].boost(boost)
@@ -122,7 +140,7 @@ class Character(AbstractCustomizable):
         Restores HP. Converts an INTEGER to a percentage. Returns the amount of HP
         healed.
         """
-        mult = 1 + self.getStatValue("potency") / 100
+        mult = 1 + self.get_stat_value("potency") / 100
         healing = self._max_hp * (float(percent) / 100) * mult
         self.remaining_hp = int(self.remaining_hp + healing)
 
@@ -135,7 +153,7 @@ class Character(AbstractCustomizable):
         """
         returns the actual amount of damage inflicted
         """
-        mult = 1 - self.getStatValue("potency") / 100
+        mult = 1 - self.get_stat_value("potency") / 100
         harming = self._max_hp * (float(percent) / 100)
         amount = int(harming * mult)
         self.take_damage(amount)
@@ -150,12 +168,12 @@ class Character(AbstractCustomizable):
         """
         Returns the amount of energy gained
         """
-        mult = 1 + self.getStatValue("potency") / 100
+        mult = 1 + self.get_stat_value("potency") / 100
         amount = int(amount * mult)
         self.energy += amount
 
-        if self.energy > self.getStatValue("energy"):
-            self.energy = self.getStatValue("energy")
+        if self.energy > self.get_stat_value("energy"):
+            self.energy = self.get_stat_value("energy")
 
         self.energy = int(self.energy)
 
@@ -168,7 +186,7 @@ class Character(AbstractCustomizable):
 
     def update(self):
         self.fire_event_listeners(UPDATE_EVENT, self)
-        self.gain_energy(self.getStatValue("energy") * 0.15)
+        self.gain_energy(self.get_stat_value("energy") * 0.15)
         for stat in self.stats.values():
             stat.update()
 
@@ -192,8 +210,11 @@ class Character(AbstractCustomizable):
             msgs.append(f'{self.name} leveled up!')
             self.xp -= self.level * 10
             self.level += 1
-            self.calcStats()
+            self._calc_stats()
             self.remaining_hp = self._max_hp
             msgs.append(self.get_display_data())
         self.xp = int(self.xp)
         return msgs
+
+    def __str__(self):
+        return self.name
